CIRC-966: Remove Joda Time.
*Summary*:
- `DateTime` -> `ZonedDateTime`
- `DateTimeZone.UTC` -> `ZoneOffset.UTC`
- `.withZone()` -> `.withZoneSameInstant()` or `.withZoneSameLocal()`
- `.toDateTime()` -> `ZonedDateTime.of()`
- `.toString(TIME_PATTERN)` -> `.format(DateTimeFormatter.ofPattern(TIME_PATTERN))`
- `ISODateTimeFormat.dateTime()` -> `DateTimeFormats.DATE_TIME`.
- `Years`, `Months`, `Weeks`, `Days`, `Hours`, `Minutes`, and `Seconds` are no longer available.
- `Interval` is no longer available, but a custom (minimal) alternative `org.folio.circulation.domain.time.Interval` is now available.
- `startDate.plus(period.timePeriod())` -> `period.plusDate(startDate)` (new methods added to `org.folio.circulation.domain.policy.Period`.
- `Period.months(duration);` -> `Duration.of(duration, ChronoUnit.MONTHS)` and similar changes for weeks, days, hours, minutes, and millis.
- `Period` -> `Duration` in some cases.
- Class-based time unit uses such as `withinSecondsAfter(seconds(2))` is replaced with integers, such as `withinSecondsAfter(2)`.
- `JsonPropertyFetcher.java` loses the `getJodaLocalDateProperty()` method as there already is a `getXXXDateProperty()` equivalent already handles `ZonedDateTime`.
- `DateTimeUtil.java` loses the `toZonedDateTime()` method as there is no longer a `DateTime` to convert to a `ZonedDateTime`.
- The `.timePeriod()` method is removed entirely from `org.folio.circulation.domain.policy.Period`.
- `DateTimeConstants.*` uses are removed and replaced with their numeric values, such as `DateTimeConstants.JANUARY` is now just `1`.

The `org.folio.circulation.domain.time.Interval` class is created to address limitations by the `Interval`, `Duration`, and/or `Period` classes in Java Time.
The goal of this is to avoid needing to add some third-party module as the use cases are believed to be sufficiently small.

As for the changes to `org.folio.circulation.domain.policy.Period`.
The Java Time has a `Period` class, but they decided to only implement *Year*, *Month*, and *Day*.
This makes the Java Time `Period` class useless here as the *Hours*, *Minutes*, *Seconds*, and *Milliseconds* are being expected and used in various places.
The `ZonedDateTime` is `final` and therefore cannot be modified.
The end result is that I reversed the logic to achieve the similar functionality with how `org.folio.circulation.domain.policy.Period` is returning `timePeriod()` (aka: a Joda Time Period).
Instead of returning this `Period`, this class accepts a `ZonedDateTime` and performs the appropriate addition or subtraction logic via the `plusDate()` and `minusDate()` methods.

In some cases, dates "between" are now calculated differently.
For example, `Days.daysBetween(X, Y)` is replaced with `Duration.between(W, Z).toDays()` such that X and Y are converted to LocalDates W and Z.

There are concerns with the ZoneId to ZoneOffset differences.
Two utility functions were created, both called: `toZoneOffset()`.
They may not be needed anymore and might be deleted or they may be needed if ZoneId must be worked with instead of ZoneOffset.
For example, the tests may need to use `ZoneId.of("America/New_York")`, rather than getting "America/New_York" as some offset like +1hour, +2 hours, etc..

At this time many of the tests are failing and I suspect that the ZoneId or ZoneOffset are somehow related.
For example, I see `[2021-04-12T08:13:44.502005-05:00[America/Chicago]]` in the errors, that `[America/Chicago]` should not be there.
The `.toString()` method on the ZonedDateTime may also be the problem.

There can be only one formatter for LocalDate.parse() and ZonedDateTime.parse().
This means that multiple try..catch.. statements must be used to traverse all of the desired valid formats.

The LocalDate.parse() and ZonedDateTime.parse() are also overly strict and refuse to handle a format that has no time (such as yyyy-MM-dd).
There is a way to build a custom formatter, but that requires setting defaults on the parser, such as:
```
    DateTimeFormatter startOfDay = new DateTimeFormatterBuilder()
      .parseCaseInsensitive()
      .append(DateTimeFormatter.ISO_LOCAL_DATE)
      .parseDefaulting(ChronoField.NANO_OF_DAY, 0)
      .parseDefaulting(ChronoField.OFFSET_SECONDS, 0)
      .toFormatter();
```

JodaTime DateTimeFormatter.dateTime() uses format 'yyyy-MM-dd'T'HH:mm:ss.SSSZZ'.
This is close DateTimeFormatter.ISO_OFFSET_DATE_TIME from Java Time but the `.SSS` is not used in DateTimeFormatter.ISO_OFFSET_DATE_TIME.
A custom formatter DateTimeFormats.DATE_TIME is implemented to provide this.

JodaTime defaults to now() when passed NULL.
JavaTime does not.
Custom static methods have been provided to normalize the dates and times where NULL may become now().

Tests like:
```
    assertThat(calculationResult.value(), is(ZonedDateTime.of(2018, 12, 31, 23, 59, 59, 0,
      UTC)));
```
can no longer be used.
The reason is that JavaTime doesn't care if the time is set to UTC.
The returned String is always in their local timezone and in their special format.
This format differs from what is expected.
The solution is to instead do the following:
```
    final ZonedDateTime desiredDueDate = ZonedDateTime.of(2018, 12, 31, 23, 59, 59, 0, UTC);

    assertThat(toDateTimeString(calculationResult.value()), is(toDateTimeString(desiredDueDate)));
```

There are also tests written with bad dates.
These bad dates now generate errors, such as:
```
java.time.DateTimeException: Invalid date 'February 29' as '2018' is not a leap year
```