using offset ids, fixed system clock and start of day
Some methods were expecting offset ids instead of the timezone string

The approach of mocking the System currentTimeMillis method was not
allowed by mockito and an alternate approach of using the jave.time
Clock class to set a fixed system clock was taken. Once the test was
finished the clock needed to be reset.

Some methods were expecting a LocalDateTime not a LocalDate and the
atStartOfDay method was used to fix this