CIRC-966: Redesign ClockManager and convert everything to use ClockManager.
Java Time does not change the system clock in the way is expected.
Instead, to utilize the system clock and then set it to some arbitrary time (be it fixed or otherwise), a custom clock must be used.
Because calling `.now()` and related always goes directly to the system clock, custom times cannot be supported.

To fix this, change everything to use a custom Clock, including "system clock" behavior.
This requires providing a globally shared Clock that everything uses.
That clock can then be set to sync with the System Clock or to some custom time (be it fixed or otherwise).

There is an existing ClockManager class that is being recycled to this end.
Everything is changed to utilize this but given the size of this change, be wary of oversights and mistakes.