CIRC-966: Be pickier about timezones.
Be more explicit when handling timezones such that `.withZoneSameInstant()` is called more.
The timezone on the ZonedDateTime may be different than the timezone being passed around as a ZoneOffset.
When obtaining these ZonedDateTimes, make sure they are in the timezone designated by the passed around ZoneOffset.

The ZoneOffset on the ShortTermLoansBaseStrategy is not being made available and is required to perform proper timezone related operations.
Change this from `private` to `protected`.

Update comments/documentation.