CIRC-966: try to fix problems and regressions.
Inconsistently has NULL where NULL needs to be passed as is or converted to date.
Implemented `formatDateTimeOptional()` to help facilitate this.
There may be places where NULL must be converted and so this change may have regression due to changes in JsonProperty

An incorrect test in OverduePeriodCalculatorService.
Should be comparing due date to system time and not due date to itself.

Do comparison against ZonedDateTime rather than sring for DueDateScheduledNoticesTests.

RenewalAPITests should be using the ClockManager.