CIRC-966: formatDateTime(), formatDateTimeOptional(), and also UTC in write().
Reviewing the original code and the current in regards to `write()` revealed that there are some cases where `now()` behavior is intended and cases where `now()` is not intended.
Update the code to use `formatDateTime()` where `now()` is intended.

The original code also was writing with dates explicitly cast to UTC.
Preserve this behavior.