CIRC-966: Use DateTimeUtil inside Interval.
Much of this functionality is duplicated.
`Interval` was created earlier on before the realization that a `DateTimeUtil` would be needed.
This changes `Interval` to better utilize `DateTimeUtil`.
Some functionality in `Interval` but not in `DateTimeUtil` is moved into `DateTimeUtil`.

The documentation comments are also stale.
Update the stale documentation comments.