CIRC-966: Latest code needs to allow bad data in tests to catch bad data.
It turns out that the normalization process that helps prevent bad data is preventing NULL cases from existing.
This results in dates being set to `now()` as per compatibility with JodaTime.

Instead, utilize "optional" behavior to allow for `null` to be passed into data as-is.
This then allows the tests to correctly fail.

Update documentation comments.