CIRC-966: Incorrect dates generated due to time zones usage.
The time zone passed to `calculateHoldShelfExpirationDate()` is not even being applied.
The behavior is now changed to set the time zone to the zone passed to `calculateHoldShelfExpirationDate()`.

Remove excess conversions.
The value to `changeHoldShelfExpirationDate()` does not need to be converted because it output write will convert the time zones.

Use the `plusDays()` and similar methods instead of using `ChronoUnit.interval()`.