CIRC-966: KeepCurrentDateStrategy fixes.
Use the utility function to generate end of day consistently.

The tests for this are also wrong.
The end of the day in UTC can be a different end of the day in a different time zone.
The tests must be comparing in the same time zone.

The JavaTime also returns a the Month Name when calling `.getMonth()`.
Change this to `.getMonthValue()`.