CIRC-966: Rollover results in next day and not current day.
In this case "Current Service Point Hours" refers not to the current hours for the service point but instead for the strategy `EndOfCurrentHoursStrategy`.
What happens is that the library is currently closed and as such there are no longer any hours available for the current day.
The next day is the next available "Current Service Point Hours".

The test incorrectly attempts to get the current day's hours rather than the next day's hours.